* fixed buildToolsVer * fix forecast response day humidity type from float to int * changed precipitation type from float to int (in mm, little to no precision lost) * changed visibility type from int to float (in km, lots of precision could've been lost) * changed formatting so that only unitSystem index is needed * added pressure, clouds, visibility, uv to info page

+++ app/src/main/java/asceapps/weatheria/data/api/ForecastResponse.kt
@@ -119,7 +119,7 @@ class ForecastResponse(
 		val condition: Condition,
 		val maxwind_kph: Float,
 		val totalprecip_mm: Float,
-		val avghumidity: Float,
+		val avghumidity: Int,
 		val avgvis_km: Float,
 		val daily_chance_of_rain: Int,
 		val daily_chance_of_snow: Int,
